pyproject.toml
[tool.poetry]
name = "lumbar-spine-classification"
version = "0.1.0"
description = ""
authors = ["remi.coscoy"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = "^2.5.1"
numpy = "^2.1.3"
PyYAML = "^6.0.2"
wandb = "^0.18.5"
torchvision = "^0.20.1"
matplotlib = "^3.9.2"
torchinfo = "^1.8.0"
tqdm = "^4.66.6"
pydicom = "^3.0.1"
kaggle = "^1.6.17"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
